Updated repository:
Have created Players, the Dealer, and started functions to
transfer cards from the Dealer to each Player.  Next stage is making
the Table and how the Players interact with the table.  The Table will
be drawn in the window - a GUI can take place around the window for the
players.  The buttons will act as inversion of control.